python-service/requirements.txt
# Python dependencies for face recognition service
# Using face-recognition library (requires dlib)
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
face-recognition>=1.3.0
git+https://github.com/ageitgey/face_recognition_models
opencv-python>=4.8.0
numpy>=1.21.0,<2.0.0
Pillow>=9.0.0
pydantic==2.5.0
requests>=2.31.0

